+++ config/jsdoc/info/conf.json
@@ -11,8 +11,9 @@
     },
     "plugins": [
         "jsdoc-plugin-typescript",
-        "config/jsdoc/info/define-plugin.cjs",
-        "config/jsdoc/info/virtual-plugin.cjs"
+        "config/jsdoc/plugins/define-plugin.cjs",
+        "config/jsdoc/plugins/virtual-plugin.cjs",
+        "config/jsdoc/plugins/default-export.cjs"
     ],
     "typescript": {
       "moduleRoot": "src"
