+++ config/jsdoc/plugins/define-plugin.cjs
