+++ config/jsdoc/plugins/virtual-plugin.cjs
